Potato-backend/requirements.txt
fastapi
uvicorn
pillow
tensorflow
numpy
pymongo
python-multipart